requirements.txt
# Big Marta — Sylvins Bot
# Dépendances Python

anthropic>=0.40.0
python-telegram-bot>=21.0
